ADDED: i_live finished

+++ corewar/include/champion.h
@@ -17,7 +17,7 @@
 
 // CHAMPION LIST
 typedef struct {
-    size_t id;
+    unsigned int id;
     char *name;
     bool is_alive;
     size_t alive_it;
